- Component: changed H-/GRow to float   -> grow >= 1 => propotion relative to other children   -> 0 < grow < 1 => percentage relative to parent   -> grow == 0 => no size change - adapted new H-/VGrow in layout panes - tweaked TestView to test new growth

+++ src/Mgx/Layout/Component.cs
@@ -8,14 +8,14 @@ using System;
 
 namespace Mgx.Layout {
     public abstract class Component : INotifyPropertyChanged {
-        private int hgrow;
-        public int HGrow {
+        private float hgrow;
+        public float HGrow {
             get {return hgrow;}
             set {SetProperty(ref hgrow, value);}
         }
 
-        private int vgrow;
-        public int VGrow {
+        private float vgrow;
+        public float VGrow {
             get {return vgrow;}
             set {SetProperty(ref vgrow, value);}
         }
